chore: move version helper methods to version file
