+++ internal/utils/utils_test.go
@@ -699,24 +699,3 @@ func TestProcessBucketSearchPage_SnapshotsNotFound(t *testing.T) {
 	assert.NotNil(t, err)
 	assert.True(t, mediaLink == "")
 }
-
-func TestRemoveCommitFromSnapshot(t *testing.T) {
-	assert.Equal(t, "elastic-agent-8.0.0-SNAPSHOT-darwin-x86_64.tar.gz", RemoveCommitFromSnapshot("elastic-agent-8.0.0-abcdef-SNAPSHOT-darwin-x86_64.tar.gz"))
-	assert.Equal(t, "8.0.0-SNAPSHOT", RemoveCommitFromSnapshot("8.0.0-a12345-SNAPSHOT"))
-	assert.Equal(t, "7.x-SNAPSHOT", RemoveCommitFromSnapshot("7.x-a12345-SNAPSHOT"))
-	assert.Equal(t, "7.14.x-SNAPSHOT", RemoveCommitFromSnapshot("7.14.x-a12345-SNAPSHOT"))
-	assert.Equal(t, "8.0.0-SNAPSHOT", RemoveCommitFromSnapshot("8.0.0-SNAPSHOT"))
-	assert.Equal(t, "7.x-SNAPSHOT", RemoveCommitFromSnapshot("7.x-SNAPSHOT"))
-	assert.Equal(t, "7.14.x-SNAPSHOT", RemoveCommitFromSnapshot("7.14.x-SNAPSHOT"))
-}
-
-func TestSnapshotHasCommit(t *testing.T) {
-	t.Run("Returns true with commits in snapshots", func(t *testing.T) {
-		assert.True(t, SnapshotHasCommit("8.0.0-a12345-SNAPSHOT"))
-	})
-
-	t.Run("Returns false with commits in snapshots", func(t *testing.T) {
-		assert.False(t, SnapshotHasCommit("7.x-SNAPSHOT"))
-		assert.False(t, SnapshotHasCommit("8.0.0-SNAPSHOT"))
-	})
-}
