+++ internal/versions_test.go
@@ -85,3 +85,24 @@ func Test_GetVersion(t *testing.T) {
 		assert.Equal(t, "1.2.3", v, "Version should not include SNAPSHOT nor commit")
 	})
 }
+
+func TestRemoveCommitFromSnapshot(t *testing.T) {
+	assert.Equal(t, "elastic-agent-8.0.0-SNAPSHOT-darwin-x86_64.tar.gz", RemoveCommitFromSnapshot("elastic-agent-8.0.0-abcdef-SNAPSHOT-darwin-x86_64.tar.gz"))
+	assert.Equal(t, "8.0.0-SNAPSHOT", RemoveCommitFromSnapshot("8.0.0-a12345-SNAPSHOT"))
+	assert.Equal(t, "7.x-SNAPSHOT", RemoveCommitFromSnapshot("7.x-a12345-SNAPSHOT"))
+	assert.Equal(t, "7.14.x-SNAPSHOT", RemoveCommitFromSnapshot("7.14.x-a12345-SNAPSHOT"))
+	assert.Equal(t, "8.0.0-SNAPSHOT", RemoveCommitFromSnapshot("8.0.0-SNAPSHOT"))
+	assert.Equal(t, "7.x-SNAPSHOT", RemoveCommitFromSnapshot("7.x-SNAPSHOT"))
+	assert.Equal(t, "7.14.x-SNAPSHOT", RemoveCommitFromSnapshot("7.14.x-SNAPSHOT"))
+}
+
+func TestSnapshotHasCommit(t *testing.T) {
+	t.Run("Returns true with commits in snapshots", func(t *testing.T) {
+		assert.True(t, SnapshotHasCommit("8.0.0-a12345-SNAPSHOT"))
+	})
+
+	t.Run("Returns false with commits in snapshots", func(t *testing.T) {
+		assert.False(t, SnapshotHasCommit("7.x-SNAPSHOT"))
+		assert.False(t, SnapshotHasCommit("8.0.0-SNAPSHOT"))
+	})
+}
